fix: build failed on linux
