+++ app/cmd/winvt_stub.go
@@ -1,0 +1,7 @@
+//go:build !windows
+// +build !windows
+
+package main
+
+func openWinVT() {
+}
